requirements.txt
uvicorn[standard]
fastapi
motor
aiosmtplib
python-dotenv
gunicorn
